requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# HRCenterNet/datasets/HanDataset.py: 4
# HRCenterNet/demo_dataset.ipynb: 5
# HRCenterNet/evaluate.py: 7
# HRCenterNet/test.py: 7
Pillow >= 7.0.0

# HRCenterNet/demo_dataset.ipynb: 2
matplotlib >= 3.3.1

# HRCenterNet/datasets/HanDataset.py: 3
# HRCenterNet/demo_dataset.ipynb: 3
# HRCenterNet/evaluate.py: 8
# HRCenterNet/test.py: 8
# HRCenterNet/utils/utility.py: 3
numpy >= 1.19.1

# HRCenterNet/demo_dataset.ipynb: 6
# HRCenterNet/utils/utility.py: 1
pandas >= 0.24.2

# HRCenterNet/evaluate.py: 9
# HRCenterNet/test.py: 9
# HRCenterNet/utils/utility.py: 2
scikit_image >= 0.15.0

# HRCenterNet/datasets/HanDataset.py: 1
# HRCenterNet/demo_dataset.ipynb: 1
# HRCenterNet/evaluate.py: 2,3
# HRCenterNet/models/HRCenterNet.py: 1,2
# HRCenterNet/models/modules.py: 1
# HRCenterNet/test.py: 2,3
# HRCenterNet/train.py: 2
# HRCenterNet/utils/losses.py: 1
# HRCenterNet/utils/utility.py: 5
torch >= 1.4.0

# HRCenterNet/datasets/HanDataset.py: 2
# HRCenterNet/demo_dataset.ipynb: 4
# HRCenterNet/evaluate.py: 4,5
# HRCenterNet/test.py: 4,5
# HRCenterNet/utils/utility.py: 4
torchvision >= 0.5.0

# HRCenterNet/train.py: 4
tqdm >= 4.31.1
